HW8: Alexey Valento

+++ src/task_8/task_8_3.py
@@ -29,7 +29,7 @@ def sin1(x: float, e: list) -> float:
     return result
 
 try:
-    x = int(input(":"))
+    x = float(input(":"))
 except ValueError:
     print("x is no digit!")
